+++ jdaviz/core/loaders/resolvers/spectral_line_db/__init__.py
@@ -1,0 +1,1 @@
+from .spectral_line_db import *  # noqa